refactor(knowledge): split the token scan out of the checker
